requirements.txt
aiogram==3.17.0
transformers==4.44.0
torch==2.4.0
ffmpeg-python==0.2.0